api #1: roles fix - roles working without errors now - enum and decorator changed directory to auth module

+++ apps/api/src/users/users.service.ts
@@ -1,10 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './entities/user.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { UserRole } from './user-role.enum';
+import { UserRole } from '../auth/enums/role.enum';
 
 @Injectable()
 export class UsersService {
@@ -45,18 +45,24 @@ export class UsersService {
     return undefined;
   }
 
-  update(id: number, updateUserDto: UpdateUserDto): Promise<User> {
-    const user: User = new User();
-    user.username = updateUserDto.username;
-    user.email = updateUserDto.email;
-    user.password = updateUserDto.password;
-    user.role = updateUserDto.role;
-    user.banStatus = updateUserDto.banStatus;
-    user.id = id;
-    return this.userRepository.save(user);
+  update(id: number, updateUserDto: UpdateUserDto, req: any): Promise<User> {
+    if (req.user.role === UserRole.Admin || req.user.userId === id) {
+      const user: User = new User();
+      user.username = updateUserDto.username;
+      user.email = updateUserDto.email;
+      user.password = updateUserDto.password;
+      user.role = updateUserDto.role;
+      user.banStatus = updateUserDto.banStatus;
+      user.id = id;
+      return this.userRepository.save(user);
+    }
+    throw new ForbiddenException();
   }
 
-  remove(id: number): Promise<{ affected?: number }> {
-    return this.userRepository.delete(id);
+  remove(id: number, req: any): Promise<{ affected?: number }> {
+    if (req.user.role === UserRole.Admin || req.user.userId === id) {
+      return this.userRepository.delete(id);
+    }
+    throw new ForbiddenException();
   }
 }
